requirements.txt
pymysql>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
snowflake-connector-python>=3.0.0
pandas>=2.0.0
